requirements.txt
pytest==8.2.2
selenium==4.23.1
webdriver-manager==4.0.2
allure-pytest==2.13.5